+++ src/test/groovy/ru/vyarus/gradle/plugin/animalsniffer/cache/AndroidCachedValidationKitTest.groovy
@@ -47,9 +47,9 @@ class AndroidCachedValidationKitTest extends AbstractKitTest {
         File file = file('/build/reports/animalsniffer/main.text')
         file.exists()
         file.readLines() == [
-                'android.Sample:9  Undefined reference (java16-sun-1.0): int Boolean.compare(boolean, boolean)',
                 'android.Sample:9  Undefined reference (android-api-level-14-4.0_r4): int Boolean.compare(boolean, boolean)',
-                'android.Sample:14  Undefined reference (android-api-level-14-4.0_r4): Object javax.naming.InitialContext.doLookup(String)'
+                'android.Sample:14  Undefined reference (android-api-level-14-4.0_r4): Object javax.naming.InitialContext.doLookup(String)',
+                'android.Sample:9  Undefined reference (java16-sun-1.0): int Boolean.compare(boolean, boolean)'
         ]
     }
 }