+++ src/test/groovy/ru/vyarus/gradle/plugin/animalsniffer/cache/CacheExcludeKitTest.groovy
@@ -31,7 +31,7 @@ class CacheExcludeKitTest extends AbstractKitTest {
             }
             
             task printSignature(type: ru.vyarus.gradle.plugin.animalsniffer.info.SignatureInfoTask) {
-                signature = animalsnifferCacheMain.outputs.files
+                signature = animalsnifferCacheMain.outputFiles
                 depth = 2
             }